signalapp/requirements.txt
pycryptodome==3.10.1

